😮 Move Caclulator class to createStatementData.js
