+++ practice/chapter01/statement.js
@@ -1,55 +1,5 @@
 import createStatementData from './createStatementData';
 
-export class PerformanceCalculator {
-  constructor(aPerformance, aPlay) {
-    this.performance = aPerformance;
-    this.play = aPlay;
-  }
-
-  get amount() {
-    throw new Error('서브클래스에서 처리하도록 설계되었습니다.');
-  }
-
-  get volumeCredits() {
-    return Math.max(this.performance.audience - 30, 0);
-  }
-}
-
-class TragedyCalculator extends PerformanceCalculator {
-  get amount() {
-    let result = 40000;
-    if (this.performance.audience > 30) {
-      result += 1000 * (this.performance.audience - 30);
-    }
-    return result;
-  }
-}
-class ComedyCalculator extends PerformanceCalculator {
-  get amount() {
-    let result = 30000;
-    if (this.performance.audience > 20) {
-      result += 10000 + 500 * (this.performance.audience - 20);
-    }
-    result += 300 * this.audience;
-    return result;
-  }
-
-  get VolumeCredits() {
-    return super.volumeCredits + Math.floor(this.performance.audience / 5);
-  }
-}
-
-export function createPerformanceCalculator(aPerformance, aPlay) {
-  switch (aPlay.type) {
-    case 'tragedy':
-      return new TragedyCalculator(aPerformance, aPlay);
-    case 'comedy':
-      return new ComedyCalculator(aPerformance, aPlay);
-    default:
-      throw new Error(`알 수 없는 장르: ${aPlay.type}`);
-  }
-}
-
 export default function statement(invoice, plays) {
   renderPlainText(createStatementData(invoice, plays));
 }
